make router a module in app
