+++ App/Client/app.js
@@ -1,63 +1,18 @@
 import Vue from 'vue';
-import VueRouter from 'vue-router';
 import VueResource from 'vue-resource'; 
-import Vuex from 'vuex';
+import VueRouter from 'vue-router';
 
-import AppTEST from './appController.js';
-import login from './Views/loginController.js';
-import video from './Views/videoController.js';
-import signup from './Views/signupController.js';
-import profile from './Views/profileController.js';
-import profileCreate from './Views/profileCreationController.js';
+import router from './Router/router.js';
 import store from './store.js';
+
+import login from './Views/loginController.js';
 
 Vue.use(VueResource);
 Vue.use(VueRouter);
-Vue.use(Vuex);
-
 
 Vue.component('login', login);
 
-
-
-
-var routes = [
-
-  {
-    path: '/',
-    component: AppTEST
-  },  
-  {
-    path: '/video',
-    component: video
-  },
-  {
-    path: '/signup',
-    component: signup
-
-  },
-  {
-    path: '/profile/:id',
-    component: profile,
-    name: 'profile'
-  },
-  {
-    path: '/profileCreate/:id',
-    component: profileCreate,
-    name: profileCreate
-  }
-];
-
-
-const router = new VueRouter({
-
-  routes
-});
-
-
-
 const app = new Vue({
-
   store,
   router
 }).$mount('.app');
